feat: Implement model loading and fix cold-start recommendations
- Implement complete model initialization (GNN, Text Encoder, Reranker)
- Fix PyTorch 2.6+ compatibility (weights_only=False for torch.load)
- Fix GATConv for bipartite graphs (add_self_loops=False)
- Add database-to-processed pipeline (db_to_processed.py)
- Fix cold-start: use fallback instead of empty results
- Add detailed logging for debugging
- Update README with complete project structure
- All services (GNN, Text Encoder, Reranker, Graph) now loaded and used

+++ src/data/graph_builder.py
@@ -56,10 +56,11 @@ class GraphBuilder:
         
         # Create edge indices: [source, target]
         # Users -> Recipes
+        # Use int64 to handle large IDs
         edge_index = torch.tensor([
-            user_indices.tolist(),
-            recipe_indices_offset.tolist()
-        ], dtype=torch.long)
+            user_indices.astype(np.int64).tolist(),
+            recipe_indices_offset.astype(np.int64).tolist()
+        ], dtype=torch.int64)
         
         # Edge attributes (ratings)
         if edge_attr and edge_attr in interactions.columns:
@@ -122,9 +123,9 @@ class GraphBuilder:
             edge_index = torch.tensor([
                 recipe_indices,
                 ingredient_indices
-            ], dtype=torch.long)
+            ], dtype=torch.int64)
         else:
-            edge_index = torch.empty((2, 0), dtype=torch.long)
+            edge_index = torch.empty((2, 0), dtype=torch.int64)
         
         logger.info(f"Built recipe-ingredient graph: {len(recipe_indices)} edges, {n_ingredients} ingredients")
